fix: :bug: Fix pipe returning 0 even on failure
