+++ src/api/util.c
@@ -1,0 +1,8 @@
+#include "minishell.h"
+
+//	frees allocated memory from shell && exits with exitcode
+void	api_exit(t_shell *shell, int exitcode)
+{
+	del_shell(shell);
+	exit(exitcode);
+}
