+++ src/prompt/prompt.c
@@ -38,20 +38,13 @@ static void	prompt_run_line(char *line, t_shell *shell)
 		return ;
 	script = parser(tokens);
 	if (!script)
-	{
-		error_malloc_msg();
-		return ;
-	}
+		return ((void)error_malloc_msg());
 	if (expander(script, shell->env) == ERR)
 		return ;
 	shell_replace_script(shell, script);
 	if (DEBUG)
 		ast_script_repr(shell->script);
-	shell_open_redirects(shell);
-	if (is_ast_command(shell->script))
-		api_exec_cmd_at(shell, 0);
-	else if (is_ast_pipeline(shell->script))
-		api_exec_pipe(shell);
+	shell_exec_script(shell);
 	shell_clear_script(shell);
 }
